Exclude old tests from workflow
