+++ setup.cfg
@@ -7,9 +7,16 @@ line_length = 88
 
 [flake8]
 extend-ignore = E501
+exclude = tests/integration/old
 
 [mypy]
 disallow_untyped_defs = true
+
+[mypy-tests.integration.old.*]
+ignore_errors = true
+
+[tool:pytest]
+norecursedirs = tests/integration/old
 
 [coverage:run]
 omit =
